Be more defensive about empty cache keys
